+++ lib/investigationCollection/index.ts
@@ -1,3 +1,5 @@
 export * from "./types";
 export * from "./dnsProvider";
 export * from "./orchestrator";
+export * from "./capabilityRegistry";
+export * from "./publicWebProvider";